requirements.txt
# Django Framework
Django

# Formatação e renderização de conteúdo
bleach
Markdown
pymdown-extensions
Pygments

# LangChain - Framework para aplicações com LLMs
langchain
langchain-core
langchain-community

# LangChain - Integrações específicas
langchain-groq
langchain-huggingface
langchain-qdrant

# LangGraph - Orquestração de agentes multi-step
langgraph

# Qdrant - Vector Database
qdrant-client

# Análise de dados
pandas
openpyxl

# Processamento de documentos PDF
PyPDF2

# HTTP Requests
requests

# Variáveis de ambiente
python-dotenv

# HuggingFace - Embeddings e modelos
sentence-transformers
# transformers 4.41+ é compatível com tokenizers>=0.19 exigido por langchain-huggingface
transformers
# Fix para conciliar langchain-huggingface (>=0.19.1) e transformers
tokenizers
torch

# FastEmbed - Embeddings rápidos para Qdrant
fastembed
